dynamicemb: relocate dtype helpers and optimizer state dim to utils/optimizer
Move DTYPE_NUM_BYTES and torch_to_dyn_emb into dynamicemb.utils, and keep get_optimizer_state_dim (plus related checkpoint-width helpers) in dynamicemb.optimizer. dynamicemb_config imports get_optimizer_state_dim only where needed for capacity / byte accounting, instead of owning those definitions. Re-export torch_to_dyn_emb from the package root via utils. Drop unused dynamicemb_config helpers (e.g. _next_power_of_2). Tests updated for import paths; checkpoint truncate/pad behavior is unchanged in intent.

Made-with: Cursor
